add image upload form, change login implementation
