+++ client/src/routes/Upload.tsx
@@ -1,0 +1,10 @@
+import ImageUploadForm from "../components/ImageUploadForm";
+
+const Upload = () => {
+
+    return (
+        <ImageUploadForm/>
+    );
+}
+
+export default Upload;
